chore(executor): extract values.rs (#116 step 3)
Pulls Django `.values()` / `.values_list()` / `.values_list_flat()`
machinery into a dedicated module:

- Per-dialect `pg/my/sqlite_cell_to_sqlvalue` decoders (private to
  values.rs since nothing else consumes them).
- `fetch_values_dict` / `fetch_values_list` / `fetch_values_flat`
  pool-level entry points.
- `MaybePgScalar` / `MaybeMyScalar` / `MaybeSqliteScalar` trait
  gates for the typed-scalar flat-fetch path.
- `bind_query_scalar_{pg,my,sqlite}` helpers.
- `impl ValuesQuerySet` / `impl ValuesListQuerySet` /
  `impl ValuesFlatQuerySet` bridge methods.

Required visibility bumps on `bind_match!`, `bind_match_mysql!`,
`bind_match_sqlite!` macros (now `pub(super)` via
`pub(super) use bind_match;`) and `bind_query` / `bind_query_my` /
`bind_query_sqlite` (now `pub(super) fn`) so siblings under
`executor/` can reach them.

mod.rs: 4490 → 4139 LOC. values.rs: 397 LOC.

1482 lib tests pass on the sqlite,tenancy litmus.
